tests/requirements.txt
numpy==1.19.1
matplotlib==3.3.0
bluepy==1.3.0
body==0.1
lib==3.0.0
paho_mqtt==1.5.0
